// src/generated/ipc/models/ListApiEndpointsRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aone IDE Tauri IPC
 * Documentation paths represent Tauri invoke commands, not network HTTP endpoints. Success payloads are serialized JSON values. Backend failures reject the invocation with a serialized error string.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ListApiEndpointsRequestRequest } from './ListApiEndpointsRequestRequest';
import {
    ListApiEndpointsRequestRequestFromJSON,
    ListApiEndpointsRequestRequestFromJSONTyped,
    ListApiEndpointsRequestRequestToJSON,
    ListApiEndpointsRequestRequestToJSONTyped,
} from './ListApiEndpointsRequestRequest';

/**
 * 
 * @export
 * @interface ListApiEndpointsRequest
 */
export interface ListApiEndpointsRequest {
    /**
     * 
     * @type {ListApiEndpointsRequestRequest}
     * @memberof ListApiEndpointsRequest
     */
    request: ListApiEndpointsRequestRequest;
}

/**
 * Check if a given object implements the ListApiEndpointsRequest interface.
 */
export function instanceOfListApiEndpointsRequest(value: object): value is ListApiEndpointsRequest {
    if (!('request' in value) || value['request'] === undefined) return false;
    return true;
}

export function ListApiEndpointsRequestFromJSON(json: any): ListApiEndpointsRequest {
    return ListApiEndpointsRequestFromJSONTyped(json, false);
}

export function ListApiEndpointsRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListApiEndpointsRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'request': ListApiEndpointsRequestRequestFromJSON(json['request']),
    };
}

export function ListApiEndpointsRequestToJSON(json: any): ListApiEndpointsRequest {
    return ListApiEndpointsRequestToJSONTyped(json, false);
}

export function ListApiEndpointsRequestToJSONTyped(value?: ListApiEndpointsRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'request': ListApiEndpointsRequestRequestToJSON(value['request']),
    };
}